Release 0.1.6: no .raw copy, int32 counts, float32 PCA and decontX layer (eca-rsi#2)
clustered.h5ad carried adata.raw, set after log1p and never diverging from
X (HVG/scale work on a copy) -- verified byte-identical on a real
per-sample output (fu2022 A2inner: data, indices, indptr). DE and marker
scoring now read X (use_raw=False); numbers unchanged, same matrix.

- integer counts wider than 4 bytes narrow to int32 in cluster_and_deg
  (guarded: integers only, max must fit; float counts untouched)
- X_pca written float32
- decontX_counts layer float32 instead of float64 -- it was the single
  largest dataset in the file
- test_clustered_output_is_slim runs cluster_and_deg end to end on an
  int64 sparse input and checks the returned object and the written file

Same rules as msp 0.3.6; ecarsi 0.2.6 applies them to organized.h5ad.
Changes file content, hence file_identity(): lands between batches only.

+++ osp/qc.py
@@ -547,7 +547,7 @@ def qc_one_sample(
         # returned DecontXResult in ourselves (same fields as its copy branch)
         ad.obs["decontX_contamination"] = res.contamination
         ad.obs["decontX_clusters"] = pd.Categorical(res.z)
-        ad.layers["decontX_counts"] = res.decontx_counts.T.tocsr()
+        ad.layers["decontX_counts"] = res.decontx_counts.T.tocsr().astype(np.float32)  # float64 doubled the layer
         still = _decontx_degenerate(ad.obs)
         if still is not None:
             # Contamination estimates are untrustworthy even with supplied
